refactor(path): change path
